Create README.md
I am Climate change enthusiastic that has been on the front burner of environment conversations. As a physical scientist who is passionate about building a sustainable future for my generation and posterity, I am promoting awareness on the subject of climate change and advocating for a just-transition, that leaves no one behind, through technology and new skills acquisitions and development.